refactor: workspace and hook
