+++ apps/electron/src/main/modules/mcp-core/mcp-manager/http/mcp-http-server.ts
@@ -161,7 +161,9 @@ export class MCPHttpServer {
           }
 
           // Get user auth token instead of workspace token
-          const { getDecryptedAuthToken } = await import("../../../auth/auth.service");
+          const { getDecryptedAuthToken } = await import(
+            "../../../auth/auth.service"
+          );
           const authToken = await getDecryptedAuthToken();
 
           // Forward the request to remote aggregator
