+++ apps/electron/src/main/modules/workflow/workflow.repository.ts
@@ -1,4 +1,4 @@
-import { getSqliteManager } from "../../core/sqlite-manager";
+import { getSqliteManager } from "../../infrastructure/database/core/sqlite-manager";
 import { WorkflowDefinition } from "@mcp_router/shared";
 import { v4 as uuidv4 } from "uuid";
